+++ mizani/scale.py
@@ -33,7 +33,7 @@ import pandas.api.types as pdtypes
 
 from .bounds import censor, rescale
 from .utils import CONTINUOUS_KINDS, DISCRETE_KINDS, min_max, match
-from .utils import multitype_sort
+from .utils import multitype_sort, get_categories
 
 
 __all__ = ['scale_continuous', 'scale_discrete']
@@ -204,13 +204,12 @@ class scale_discrete:
 
         # Train i.e. get the new values
         if pdtypes.is_categorical_dtype(new_data):
-            try:
-                new = list(new_data.cat.categories)  # series
-            except AttributeError:
-                new = list(new_data.categories)      # plain categorical
+            categories = get_categories(new_data)
             if drop:
                 present = set(new_data.drop_duplicates())
-                new = [i for i in new if i in present]
+                new = [i for i in categories if i in present]
+            else:
+                new = list(categories)
         else:
             try:
                 new = np.unique(new_data)
@@ -220,13 +219,19 @@ class scale_discrete:
                 new = list(set(new_data))
                 new = multitype_sort(new)
 
-        # Add nan if required
-        if has_na and not na_rm:
-            new = np.hstack([new, np.nan])
-
         # update old
         old_set = set(old)
-        return list(old) + [i for i in new if (i not in old_set)]
+        if pdtypes.is_categorical_dtype(new_data):
+            # The limits are in the order of the categories
+            final = old_set | set(new)
+            limits = [c for c in categories if c in final]
+        else:
+            limits = list(old) + [i for i in new if (i not in old_set)]
+
+        # Add nan if required
+        if has_na and not na_rm:
+            limits.append(np.nan)
+        return limits
 
     @classmethod
     def map(cls, x, palette, limits, na_value=None):