+++ mizani/tests/test_scale.py
@@ -39,6 +39,9 @@ def test_scale_discrete():
     def assert_equal_with_nan(lst1, lst2):
         assert lst1[:-1] == lst2[:-1] and np.isnan(lst2[-1])
 
+    def SCategorical(*args, **kwargs):
+        return pd.Series(pd.Categorical(*args, **kwargs))
+
     x = ['a', 'b', 'c', 'a']
     # apply
     scaled = scale_discrete.apply(x, np.arange)
@@ -87,3 +90,11 @@ def test_scale_discrete():
 
     limits = scale_discrete.train(x[:2], drop=False)
     assert limits == ['a', 'b', 'c']
+
+    # Disrete Scale training maintains order of categoricals
+    cats = ['x0', 'x1', 'x2', 'x3', 'x4']
+    s1 = SCategorical(['x1', 'x2'], categories=cats)
+    s2 = SCategorical(['x0', 'x2'], categories=cats)
+    limits = scale_discrete.train(s1, drop=True)
+    limits = scale_discrete.train(s2, limits, drop=True)
+    assert limits == ['x0', 'x1', 'x2']